post-meeting changes and notes added
The opening value is the result of a random process, provided either as a string input to the commit function, or is the result of random coin flips in the commit function. No need for opening value to persist elsewhere, other than in the binding experiment, where the adversary must choose two opening values to test against.

Note sure if this is yet the right approach. Will revisit perfect binding definitions and reexamine.